requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0
requests>=2.31
cachetools>=5.3
orjson>=3.9
